[REFACTOR] - Fixed push to cloneAndBuildWin according to issue #74 (#79)
* [REFACTOR] - Added push capabillities according to issue #74 
Fixed so that the server works also on push not just pull.
Added json string for push and a if statement in cloneAndBuildWin so
that the json string are either push or pull depending on wich ever is
recived by server.

* Added testclass.java to commit

* [REFACTOR] - Fixed push to cloneAndBuildWin according to issue #74
Fixed push to cloneAndBUildWin and added the function to run scripts in
CIServer. Also added a lock so that different requests wait untill
another is finished. Also added some futher documentation to the methods
to reflect the changes. New string input was added to cloneAndBuildWin
to make it easy to determine which request was asked. Futher testing
when log site is up and running. close #74